Refactor open orders lists
