+++ components/Orders/OpenOrdersTab.tsx
@@ -1,0 +1,126 @@
+import { useCallback, useState } from 'react';
+import { Stack, Table, Button, Group, Text } from '@mantine/core';
+import { useWallet } from '@solana/wallet-adapter-react';
+import { Transaction } from '@solana/web3.js';
+import { BN } from '@coral-xyz/anchor';
+import { OpenOrdersAccountWithKey } from '@/lib/types';
+import { useOpenbookTwap } from '@/hooks/useOpenbookTwap';
+import { useTransactionSender } from '@/hooks/useTransactionSender';
+import { useProposal } from '@/contexts/ProposalContext';
+import { isPartiallyFilled } from '@/lib/openbook';
+import { OpenOrderRow } from './OpenOrderRow';
+
+const headers = ['Order ID', 'Market', 'Status', 'Size', 'Price', 'Notional', 'Actions'];
+
+export function OpenOrdersTab({ orders }: { orders: OpenOrdersAccountWithKey[] }) {
+  const { markets, isCranking, crankMarkets } = useProposal();
+  const sender = useTransactionSender();
+  const wallet = useWallet();
+  const { fetchOpenOrders, proposal } = useProposal();
+  const { cancelOrderTransactions, settleFundsTransactions } = useOpenbookTwap();
+
+  const [isCanceling, setIsCanceling] = useState<boolean>(false);
+  const [isSettling, setIsSettling] = useState<boolean>(false);
+
+  const handleCancelAll = useCallback(async () => {
+    if (!proposal || !markets) return;
+
+    const txs = (
+      await Promise.all(
+        orders.map((order) =>
+          cancelOrderTransactions(
+            new BN(order.account.accountNum),
+            proposal.account.openbookPassMarket.equals(order.account.market)
+              ? { publicKey: proposal.account.openbookPassMarket, account: markets.pass }
+              : { publicKey: proposal.account.openbookFailMarket, account: markets.fail },
+          ),
+        ),
+      )
+    )
+      .flat()
+      .filter(Boolean);
+
+    if (!wallet.publicKey || !txs) return;
+
+    try {
+      setIsCanceling(true);
+      // Filtered undefined already
+      await sender.send(txs as Transaction[]);
+      // We already return above if the wallet doesn't have a public key
+      await fetchOpenOrders(wallet.publicKey!);
+    } catch (err) {
+      console.error(err);
+    } finally {
+      setIsCanceling(false);
+    }
+  }, [proposal, markets, wallet.publicKey, cancelOrderTransactions, fetchOpenOrders, sender]);
+
+  const handleSettleAllFunds = useCallback(async () => {
+    if (!proposal || !markets) return;
+
+    setIsSettling(true);
+    try {
+      // HACK: Assumes all orders are for the same market
+      const pass = orders[0].account.market.equals(proposal.account.openbookPassMarket);
+      const txs = (
+        await Promise.all(
+          orders
+            .filter((order) => isPartiallyFilled(order))
+            .map((order) =>
+              settleFundsTransactions(
+                order.account.accountNum,
+                pass,
+                proposal,
+                pass
+                  ? { account: markets.pass, publicKey: proposal.account.openbookPassMarket }
+                  : { account: markets.fail, publicKey: proposal.account.openbookFailMarket },
+              ),
+            ),
+        )
+      )
+        .flat()
+        .filter(Boolean);
+
+      if (!txs) return;
+      sender.send(txs as Transaction[]);
+    } finally {
+      setIsSettling(false);
+    }
+  }, [orders, proposal, settleFundsTransactions]);
+
+  return (
+    <Stack>
+      <Text>
+        If you see orders here with a settle button, you can settle them to redeem the partial fill
+        amount. These exist when there is a balance available within the Open Orders Account.
+      </Text>
+      <Group justify="space-around">
+        <Button loading={isCranking} onClick={() => crankMarkets()}>
+          Crank 🐷
+        </Button>
+        <Button color="red" loading={isCanceling} onClick={handleCancelAll}>
+          Cancel all orders
+        </Button>
+        <Button loading={isSettling} onClick={handleSettleAllFunds}>
+          Settle {orders.filter((order) => isPartiallyFilled(order)).length} orders
+        </Button>
+      </Group>
+      <Table>
+        <Table.Thead>
+          <Table.Tr>
+            {headers.map((header) => (
+              <Table.Th key={header}>{header}</Table.Th>
+            ))}
+          </Table.Tr>
+        </Table.Thead>
+        <Table.Tbody>
+          {orders && orders.length > 0 ? (
+            orders.map((order) => <OpenOrderRow order={order} />)
+          ) : (
+            <Table.Tr>No Orders Found</Table.Tr>
+          )}
+        </Table.Tbody>
+      </Table>
+    </Stack>
+  );
+}
