+++ contexts/ProposalContext.tsx
@@ -1,6 +1,6 @@
 import { createContext, useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import { useConnection, useWallet } from '@solana/wallet-adapter-react';
-import { PublicKey, Transaction } from '@solana/web3.js';
+import { PublicKey, Transaction, VersionedTransaction } from '@solana/web3.js';
 import {
   getAssociatedTokenAddressSync,
   createAssociatedTokenAccountInstruction,
@@ -12,12 +12,12 @@ import {
   OpenOrdersAccountWithKey,
   OrderBook,
   ProposalAccountWithKey,
+  LeafNode,
 } from '@/lib/types';
 import { useAutocrat } from '@/contexts/AutocratContext';
 import { useConditionalVault } from '@/hooks/useConditionalVault';
 import { useOpenbookTwap } from '@/hooks/useOpenbookTwap';
 import { useTransactionSender } from '@/hooks/useTransactionSender';
-import { LeafNode } from '@/lib/types';
 import { getLeafNodes } from '../lib/openbook';
 import { debounce } from '../lib/utils';
 
@@ -31,7 +31,7 @@ export interface ProposalInterface {
   isCranking: boolean;
   metaDisabled: boolean;
   usdcDisabled: boolean;
-  handleCrank: (isPassMarket: boolean, individualEvent?: PublicKey) => Promise<void>;
+  crankMarkets: (individualEvent?: PublicKey) => Promise<void>;
   fetchOpenOrders: (owner: PublicKey) => Promise<void>;
   fetchMarketsInfo: () => Promise<void>;
   createTokenAccounts: (fromBase?: boolean) => Promise<void>;
@@ -492,23 +492,27 @@ export function ProposalProvider({
     ],
   );
 
-  const handleCrank = useCallback(
-    async (isPassMarket: boolean, individualEvent?: PublicKey) => {
+  const crankMarkets = useCallback(
+    async (individualEvent?: PublicKey) => {
       if (!proposal || !markets || !wallet?.publicKey) return;
-      let marketAccounts: MarketAccountWithKey = {
-        publicKey: markets.passTwap.market,
-        account: markets.pass,
-      };
-      let { eventHeap } = markets.pass;
-      if (!isPassMarket) {
-        marketAccounts = { publicKey: markets.failTwap.market, account: markets.fail };
-        eventHeap = markets.fail.eventHeap;
-      }
       try {
         setIsCranking(true);
-        const txs = await crankMarketTransactions(marketAccounts, eventHeap, individualEvent);
-        if (!txs) return;
-        await sender.send(txs);
+        const passTxs = await crankMarketTransactions(
+          {
+            publicKey: markets.passTwap.market,
+            account: markets.pass,
+          },
+          markets.pass.eventHeap,
+          individualEvent,
+        );
+        const failTxs = await crankMarketTransactions(
+          { publicKey: markets.failTwap.market, account: markets.fail },
+          markets.fail.eventHeap,
+          individualEvent,
+        );
+        if (!passTxs || !failTxs) return;
+        const txs = [...passTxs, ...failTxs].filter(Boolean);
+        await sender.send(txs as VersionedTransaction[]);
         fetchOpenOrders(wallet.publicKey);
       } catch (err) {
         console.error(err);
@@ -535,7 +539,7 @@ export function ProposalProvider({
         fetchMarketsInfo,
         createTokenAccounts,
         createTokenAccountsTransactions,
-        handleCrank,
+        crankMarkets,
         finalizeProposalTransactions,
         mintTokensTransactions,
         mintTokens,
